Extract getFirstIlmo to db-submodule
